restructure samplers as strategies
