+++ bofire/data_models/api.py
@@ -21,8 +21,14 @@ from bofire.data_models.features.api import (  # noqa: F401
 from bofire.data_models.kernels.api import AnyKernel, Kernel  # noqa: F401
 from bofire.data_models.objectives.api import AnyObjective, Objective  # noqa: F401
 from bofire.data_models.priors.api import AnyPrior, Prior  # noqa: F401
-from bofire.data_models.samplers.api import AnySampler, Sampler  # noqa: F401
-from bofire.data_models.strategies.api import AnyStrategy, Strategy  # noqa: F401
+from bofire.data_models.strategies.api import (  # noqa: F401
+    AnyPredictive,
+    AnySampler,
+    AnyStrategy,
+    PredictiveStrategy,
+    SamplerStrategy,
+    Strategy,
+)
 from bofire.data_models.surrogates.api import (  # noqa: F401
     AnyBotorchSurrogate,
     AnySurrogate,
@@ -40,7 +46,6 @@ AnyThing = [
         AnySurrogate,
         AnyObjective,
         AnyPrior,
-        AnySampler,
         AnyStrategy,
         Domain,
     ]
