+++ bofire/data_models/strategies/samplers/sampler.py
@@ -1,0 +1,7 @@
+from bofire.data_models.strategies.strategy import Strategy
+
+
+class SamplerStrategy(Strategy):
+    """Base class for sampling methods in BoFire for sampling from constrained input spaces."""
+
+    pass
