+++ bofire/strategies/api.py
@@ -1,14 +1,16 @@
-import bofire.data_models.strategies.api as data_models  # noqa: F401
-from bofire.strategies.botorch import BotorchStrategy  # noqa: F401
 from bofire.strategies.mapper import map  # noqa: F401
-from bofire.strategies.predictive import PredictiveStrategy  # noqa: F401
-from bofire.strategies.qehvi import QehviStrategy  # noqa: F401
-from bofire.strategies.qnehvi import QnehviStrategy  # noqa: F401
-from bofire.strategies.qparego import QparegoStrategy  # noqa: F401
-from bofire.strategies.random import RandomStrategy  # noqa: F401
-from bofire.strategies.sobo import (  # noqa: F401
+from bofire.strategies.predictives.botorch import BotorchStrategy  # noqa: F401
+from bofire.strategies.predictives.predictive import PredictiveStrategy  # noqa: F401
+from bofire.strategies.predictives.qehvi import QehviStrategy  # noqa: F401
+from bofire.strategies.predictives.qnehvi import QnehviStrategy  # noqa: F401
+from bofire.strategies.predictives.qparego import QparegoStrategy  # noqa: F401
+from bofire.strategies.predictives.sobo import (  # noqa: F401
     AdditiveSoboStrategy,
     MultiplicativeSoboStrategy,
     SoboStrategy,
 )
+from bofire.strategies.random import RandomStrategy  # noqa: F401
+from bofire.strategies.samplers.polytope import PolytopeSampler  # noqa: F401
+from bofire.strategies.samplers.rejection import RejectionSampler  # noqa: F401
+from bofire.strategies.samplers.sampler import SamplerStrategy  # noqa: F401
 from bofire.strategies.strategy import Strategy  # noqa: F401
