+++ bofire/strategies/data_models/values.py
@@ -7,6 +7,16 @@ from bofire.data_models.base import BaseModel
 Value = Union[float, str, int]
 
 
+class InputValue(BaseModel):
+    """Bofire input value.
+
+    Attributes:
+        value (Union[float, str, int]): The input value.
+    """
+
+    value: Value
+
+
 class OutputValue(BaseModel):
     """Bofire predicted output value.
 
@@ -19,13 +29,3 @@ class OutputValue(BaseModel):
     predictedValue: Value
     standardDeviation: Annotated[float, Field(ge=0)]
     objective: float
-
-
-class InputValue(BaseModel):
-    """Bofire input value.
-
-    Attributes:
-        value (Union[float, str, int]): The input value.
-    """
-
-    value: Value
