+++ bofire/strategies/mapper.py
@@ -1,16 +1,21 @@
 from typing import Dict, Type
 
 import bofire.data_models.strategies.api as data_models
-from bofire.strategies.qehvi import QehviStrategy
-from bofire.strategies.qnehvi import QnehviStrategy
-from bofire.strategies.qparego import QparegoStrategy
-from bofire.strategies.random import RandomStrategy
-from bofire.strategies.sobo import (
+from bofire.strategies.predictives.botorch import BotorchStrategy  # noqa: F401
+from bofire.strategies.predictives.predictive import PredictiveStrategy  # noqa: F401
+from bofire.strategies.predictives.qehvi import QehviStrategy  # noqa: F401
+from bofire.strategies.predictives.qnehvi import QnehviStrategy  # noqa: F401
+from bofire.strategies.predictives.qparego import QparegoStrategy  # noqa: F401
+from bofire.strategies.predictives.sobo import (  # noqa: F401
     AdditiveSoboStrategy,
     MultiplicativeSoboStrategy,
     SoboStrategy,
 )
-from bofire.strategies.strategy import Strategy
+from bofire.strategies.random import RandomStrategy  # noqa: F401
+from bofire.strategies.samplers.polytope import PolytopeSampler  # noqa: F401
+from bofire.strategies.samplers.rejection import RejectionSampler  # noqa: F401
+from bofire.strategies.samplers.sampler import SamplerStrategy  # noqa: F401
+from bofire.strategies.strategy import Strategy  # noqa: F401
 
 STRATEGY_MAP: Dict[Type[data_models.Strategy], Type[Strategy]] = {
     data_models.RandomStrategy: RandomStrategy,
@@ -20,6 +25,8 @@ STRATEGY_MAP: Dict[Type[data_models.Strategy], Type[Strategy]] = {
     data_models.QehviStrategy: QehviStrategy,
     data_models.QnehviStrategy: QnehviStrategy,
     data_models.QparegoStrategy: QparegoStrategy,
+    data_models.PolytopeSampler: PolytopeSampler,
+    data_models.RejectionSampler: RejectionSampler,
 }
