+++ bofire/strategies/multiobjective.py
@@ -1,3 +1,5 @@
+# TODO: move this to bofire.utils
+
 from typing import Optional
 
 import numpy as np
