+++ bofire/strategies/predictives/qehvi.py
@@ -15,8 +15,8 @@ from botorch.utils.multi_objective.box_decompositions.non_dominated import (
 
 from bofire.data_models.objectives.api import BotorchConstrainedObjective
 from bofire.data_models.strategies.api import QehviStrategy as DataModel
-from bofire.strategies.botorch import BotorchStrategy
 from bofire.strategies.multiobjective import get_ref_point_mask
+from bofire.strategies.predictives.botorch import BotorchStrategy
 from bofire.surrogates.torch_tools import tkwargs
