+++ bofire/strategies/random.py
@@ -4,9 +4,9 @@ import pandas as pd
 from pydantic.error_wrappers import ValidationError
 from pydantic.types import PositiveInt
 
-from bofire.data_models.api import AnySampler
-from bofire.data_models.samplers.api import PolytopeSampler, RejectionSampler
-from bofire.data_models.strategies.api import RandomStrategy as DataModel
+import bofire.data_models.strategies.api as data_models
+from bofire.strategies.samplers.polytope import PolytopeSampler
+from bofire.strategies.samplers.rejection import RejectionSampler
 from bofire.strategies.strategy import Strategy
 
 
@@ -19,19 +19,21 @@ class RandomStrategy(Strategy):
 
     def __init__(
         self,
-        data_model: DataModel,
+        data_model: data_models.RandomStrategy,
         **kwargs,
     ):
         super().__init__(data_model=data_model, **kwargs)
         self._init_sampler()
 
-    sampler: Optional[AnySampler] = None
-
     def _init_sampler(self) -> None:
         errors = []
-        for S in [PolytopeSampler, RejectionSampler]:
+        for dm, functional in [
+            (data_models.PolytopeSampler, PolytopeSampler),
+            (data_models.RejectionSampler, RejectionSampler),
+        ]:
             try:
-                self.sampler = S(domain=self.domain)
+                data_model = dm(domain=self.domain)
+                self.sampler = functional(data_model=data_model)
                 break
             except ValidationError as err:
                 errors.append(err)
