+++ tests/bofire/data_models/specs/strategies.py
@@ -1,6 +1,8 @@
 import bofire.data_models.strategies.api as strategies
 from bofire.data_models.acquisition_functions.api import qPI
-from bofire.data_models.enum import CategoricalMethodEnum
+from bofire.data_models.domain.api import Domain, Inputs
+from bofire.data_models.enum import CategoricalMethodEnum, SamplingMethodEnum
+from bofire.data_models.features.api import ContinuousInput
 from tests.bofire.data_models.specs.api import domain
 from tests.bofire.data_models.specs.specs import Specs
 
@@ -69,5 +71,38 @@ specs.add_valid(
     strategies.RandomStrategy,
     lambda: {
         "domain": domain.valid().obj().dict(),
+    },
+)
+
+specs.add_valid(
+    strategies.PolytopeSampler,
+    lambda: {
+        "domain": Domain(
+            input_features=Inputs(
+                features=[
+                    ContinuousInput(key=f"x_{i}", lower_bound=0, upper_bound=1)
+                    for i in range(2)
+                ]
+            ),
+        ),
+        "fallback_sampling_method": SamplingMethodEnum.UNIFORM,
+    },
+)
+specs.add_valid(
+    strategies.RejectionSampler,
+    lambda: {
+        "domain": Domain(
+            input_features=Inputs(
+                features=[
+                    ContinuousInput(key=f"x_{i}", lower_bound=0, upper_bound=1)
+                    for i in range(2)
+                ]
+            )
+        ),
+        "max_iters": 1000,
+        "num_base_samples": 1000,
+        "sampling_method": SamplingMethodEnum.UNIFORM,
+        "num_base_samples": 1000,
+        "max_iters": 1000,
     },
 )
