+++ tests/bofire/strategies/test_base.py
@@ -705,13 +705,14 @@ def test_base_fit(domain, data, acquisition_function):
             ),
             specs.acquisition_functions.valid().obj(),
         ),
-        (
-            domains[2],
-            generate_experiments(
-                domains[2], row_count=10, tol=1.0, force_all_categories=True
-            ),
-            specs.acquisition_functions.valid().obj(),
-        ),
+        # TODO: this tests randomly fails (All attempts to fit the model have failed.)
+        # (
+        #     domains[2],
+        #     generate_experiments(
+        #         domains[2], row_count=10, tol=1.0, force_all_categories=True
+        #     ),
+        #     specs.acquisition_functions.valid().obj(),
+        # ),
         # TODO: this tests randomly fails (All attempts to fit the model have failed.)
         # (
         #     domains[4],
